# studentinfo_v5/docs/swagger.yaml
basePath: /
definitions:
  main.Student:
    properties:
      address:
        type: string
      bloodgroup:
        type: string
      contact:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample student information system server.
  title: Student Information System API
  version: "1.0"
paths:
  /:
    get:
      consumes:
      - application/json
      description: API welcome message
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: API Root
      tags:
      - home
  /create:
    post:
      consumes:
      - application/json
      description: create a new student
      parameters:
      - description: Student Data
        in: body
        name: student
        required: true
        schema:
          $ref: '#/definitions/main.Student'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new student
      tags:
      - students
  /list:
    get:
      consumes:
      - application/json
      description: get the list of all students
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Student'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all students
      tags:
      - students
swagger: "2.0"