# studentinfo_v5/main.go
package main

import (
	"encoding/json"
	"log"
	"net/http"

	_ "expense/studentinfo_v5/docs" // docs is generated by Swag CLI, you have to import it.

	httpSwagger "github.com/swaggo/http-swagger"
)

// @title Student Information System API
// @version 1.0
// @description This is a sample student information system server.
// @host localhost:8080
// @BasePath /

func main() {
	http.HandleFunc("/", homeHandler)
	http.HandleFunc("/list", listHandler)
	http.HandleFunc("/create", createHandler)
	http.HandleFunc("/swagger/", httpSwagger.WrapHandler)

	log.Println("Server started on http://localhost:8080")
	if err := http.ListenAndServe(":8080", nil); err != nil {
		log.Fatal(err)
	}
}

// homeHandler godoc
// @Summary API Root
// @Description API welcome message
// @Tags home
// @Accept  json
// @Produce  json
// @Success 200 {object} map[string]string
// @Router / [get]
func homeHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]string{"message": "Welcome to Student Information System API"})
}

// listHandler godoc
// @Summary Get all students
// @Description get the list of all students
// @Tags students
// @Accept  json
// @Produce  json
// @Success 200 {array} Student
// @Failure 500 {object} map[string]string
// @Router /list [get]
func listHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	repo := StudentRepo{}
	students, err := repo.getStudents()
	if err != nil {
		log.Println(err)
		w.WriteHeader(http.StatusInternalServerError)
		json.NewEncoder(w).Encode(map[string]string{"error": "Internal Server Error"})
		return
	}
	json.NewEncoder(w).Encode(students)
}

// createHandler godoc
// @Summary Create a new student
// @Description create a new student
// @Tags students
// @Accept  json
// @Produce  json
// @Param student body Student true "Student Data"
// @Success 201 {object} map[string]string
// @Failure 400 {object} map[string]string
// @Failure 405 {object} map[string]string
// @Failure 500 {object} map[string]string
// @Router /create [post]
func createHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")

	if r.Method != "POST" {
		w.WriteHeader(http.StatusMethodNotAllowed)
		json.NewEncoder(w).Encode(map[string]string{"error": "Method Not Allowed"})
		return
	}

	var student Student
	if err := json.NewDecoder(r.Body).Decode(&student); err != nil {
		w.WriteHeader(http.StatusBadRequest)
		json.NewEncoder(w).Encode(map[string]string{"error": "Invalid request payload"})
		return
	}

	repo := StudentRepo{}
	if err := repo.insertStudent(student); err != nil {
		log.Println(err)
		w.WriteHeader(http.StatusInternalServerError)
		json.NewEncoder(w).Encode(map[string]string{"error": "Internal Server Error"})
		return
	}

	w.WriteHeader(http.StatusCreated)
	json.NewEncoder(w).Encode(map[string]string{"message": "Student inserted successfully"})
}